fix: address workspace chat tab review
