+++ .changeset/workspace-chat-tabs.md
@@ -1,0 +1,6 @@
+---
+"@open-codesign/desktop": patch
+"@open-codesign/i18n": patch
+---
+
+Add workspace chat tabs for starting and switching fresh conversations that share the same workspace.
